refactor: reduce duplicated feature flag code
